refactor: reorganize code significantly, fix timeout printing error
