+++ README.bazel.md
@@ -7,7 +7,7 @@
 ## External Depedencies
 This Rust project doesn't use Cargo, instead specifying dependencies through Bazel. To add a dependency, similar to how you would normally add a dependency in Cargo.toml, specify dependencies in [MODULE.bazel](MODULE.bazel) using `crate.spec(package = "my_package", version = "1.2.3")`. Then, include them in the `deps` argument of your Rust target like `"@crates//:package_name"`.
 
-## Usage - Main Server
+## Usage
 There are 4 ways we would want Bazel to build/run our project:
 1. Next.js-only dev
 3. Rust-only dev
@@ -51,13 +51,3 @@ To build or run the final product, run:
 `bazel run --config=release //backend:prod`  
 
 The final product would be the binary found in `bazel-bin/backend/prod`.
-
-## Usage - Discover Cameras
-To build the auxiliary binary for discovering cameras on the LAN, run the following command:  
-`bazel build //backend:discover_cameras`  
-
-## Usage - Record from Cameras
-To run the auxiliary binary for recording from the discovered cameras on the LAN, run the following command:
-```
-bazel run //backend:record_from_cameras -- --camera "Lucid Vision Labs-ATP124S-M-224300917" --resolution 4k --exposure-us 25.4 --frame-rate-hz 30 --save-recordings-dir ~/Downloads/camera_out --max-frames 100
-```
