+++ backend/src/camera/mod.rs
@@ -1,0 +1,3 @@
+pub mod aravis_utils;
+pub mod discovery;
+pub mod record;
