+++ backend/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod camera;
 pub mod camera_ingest;
 pub mod computer_vision;
 pub mod pipeline;
